requirements.txt
absl-py==0.1.10
astroid==1.6.0
attrs==17.4.0
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
click==6.7
cliff==2.8.0
cmd2==0.7.9
configparser==3.5.0
cssselect==1.0.3
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
fire==0.1.2
flickrapi==2.4.0
futures==3.1.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
invoke==0.22.0
isort==4.2.15
joblib==0.11
jpeg4py==0.1.2
kaggle-cli==0.12.13
Keras==2.1.3
-e git+https://www.github.com/keras-team/keras-contrib.git@17d89532ba086f17b15e7b918f6bdf3837703b56#egg=keras_contrib
lazy-object-proxy==1.3.1
lightgbm==2.0.12
lxml==4.0.0
Markdown==2.6.11
matplotlib==2.1.1
mccabe==0.6.1
MechanicalSoup==0.8.0
networkx==2.0
numpy==1.14.0
oauth==1.0.1
oauthlib==2.0.6
opencv-contrib-python==3.4.0.12
pandas==0.22.0
pbr==3.1.1
Pillow==5.0.0
pluggy==0.6.0
prettytable==0.7.2
progressbar2==3.34.3
protobuf==3.5.1
py==1.5.2
pycparser==2.18
pylint==1.8.1
pyparsing==2.2.0
pyperclip==1.6.0
pytest==3.3.2
python-dateutil==2.6.1
python-dotenv==0.7.1
python-utils==2.2.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
retry==0.9.2
rope==0.10.7
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
shell==1.0.1
six==1.11.0
stevedore==1.28.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
tinydb==3.7.0
tqdm==4.19.5
urllib3==1.22
Werkzeug==0.14.1
wget==3.2
wrapt==1.10.11
